+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - New dataloader for classification metrics (precision, recall, f1, support)
+- Standard deviation for run latency in experiments table
 
 ### Changed
 - Updated the `ProjectPageHeader` component to display classification metrics